// FrontEnd/src/app/services/api-backend/model/caso.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Paciente } from './paciente';
import { Cita } from './cita';
import { AsignacionesCama } from './asignacionesCama';
import { DiagnosticosCaso } from './diagnosticosCaso';


export interface Caso { 
    idCaso?: number;
    idPaciente?: number;
    fechaInicio?: string | null;
    fechaFin?: string | null;
    motivoFinalizacion?: string | null;
    idPacienteNavigation?: Paciente;
    asignacionesCamas?: Array<AsignacionesCama> | null;
    cita?: Array<Cita> | null;
    diagnosticosCasos?: Array<DiagnosticosCaso> | null;
}