// FrontEnd/src/app/services/api-backend/model/paciente.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Municipio } from './municipio';
import { HistoriasClinica } from './historiasClinica';
import { Caso } from './caso';


export interface Paciente { 
    idPaciente?: number;
    nombres?: string | null;
    apellidos?: string | null;
    dpi?: string | null;
    fechaNacimiento?: string;
    direccion?: string | null;
    fotografia?: string | null;
    idMunicipio?: number;
    idMunicipioNavigation?: Municipio;
    casos?: Array<Caso> | null;
    historiasClinicas?: Array<HistoriasClinica> | null;
}